+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/logic/service/MaterialPartnerRelationService.java
@@ -59,6 +59,10 @@ public interface MaterialPartnerRelationService {
 
     List<Material> findAllByPartnerMaterialNumber(String partnerMaterialNumber);
 
+    List<MaterialPartnerRelation> findAllByOwnMaterialNumber(String ownMaterialNumber);
+
+    List<MaterialPartnerRelation> findAllSupplierRelations(String ownMaterialNumber);
+
     boolean partnerSuppliesMaterial(Material material, Partner partner);
 
     boolean partnerOrdersProduct(Material material, Partner partner);