+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/logic/service/MaterialPartnerRelationServiceImpl.java
@@ -554,6 +554,22 @@ public class MaterialPartnerRelationServiceImpl implements MaterialPartnerRelati
     }
 
     /**
+     * Returns a list of all MaterialPartnerRelations for the given own material number.
+     *
+     * @param ownMaterialNumber the own material number to search for
+     * @return a list of MaterialPartnerRelations for this material
+     */
+    @Override
+    public List<MaterialPartnerRelation> findAllByOwnMaterialNumber(String ownMaterialNumber) {
+        return mprRepository.findAllByMaterial_OwnMaterialNumber(ownMaterialNumber);
+    }
+
+    @Override
+    public List<MaterialPartnerRelation> findAllSupplierRelations(String ownMaterialNumber) {
+        return mprRepository.findAllByMaterial_OwnMaterialNumberAndPartnerSuppliesMaterialIsTrue(ownMaterialNumber);
+    }
+
+    /**
      * Returns a list of all Materials, for which a MaterialPartnerRelation exists,
      * where the partner is using the given partnerMaterialNumber.
      *