+++ frontend/src/models/constants/config.ts
@@ -46,6 +46,7 @@ const app = {
     ENDPOINT_ALL_PARTNERS: import.meta.env.VITE_ENDPOINT_ALL_PARTNERS.trim() as string,
     ENDPOINT_MATERIAL_PARTNER_RELATIONS: import.meta.env.VITE_ENDPOINT_MATERIAL_PARTNER_RELATIONS.trim() as string,
     ENDPOINT_PARTNER_DATA_UPDATE_BATCH: import.meta.env.VITE_ENDPOINT_PARTNER_DATA_UPDATE_BATCH.trim() as string,
+    ENDPOINT_MATERIAL_RELATIONS: import.meta.env.VITE_ENDPOINT_MATERIAL_RELATIONS.trim() as string,
 };
 
 const auth = {